agent/requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
cryptography==43.0.3
PyYAML==6.0.2
